connecting components to state
calling the clothing API in App component to set the initial state of the app.
displaying the available clothes according to type inside CreatingSets component.
creating 3 sections inside Home component: 1 with a link for navigating to saved sets, 2 for displaying amount of every type of clothing and 3 with 3 buttons to pick new type of clothing.
creating 2 more reducers, 1 for handling saved sets and the other for handling the available clothing to pick from