creating saved sets page and its logic handling
combining actions in the reducer for better performance.
creating saved sets page, displaying the sets with the time it took to complete them, a delete button with a modal and a message displaying if no sets were picked.
creating add new set logic to handle adding new set to the store.